refactor: service, serviceImpl로 파일 나누기, 서치 utils 이용하여 다시 구현
